+++ include/qtpokit/pokitpro.h
@@ -11,8 +11,6 @@
 
 #include "qtpokit_global.h"
 
-#include <QVariant>
-
 QTPOKIT_BEGIN_NAMESPACE
 
 /// Encapsulates details specific to Pokit Pro devices.
@@ -26,7 +24,7 @@ namespace PokitPro {
         AutoRange = 255 ///< Auto-range.
     };
     QTPOKIT_EXPORT QString toString(const CapacitanceRange &range);
-    QTPOKIT_EXPORT QVariant maxValue(const CapacitanceRange &range);
+    QTPOKIT_EXPORT quint32 maxValue(const CapacitanceRange &range);
 
 
     /// Values supported by the Pokit Pro's `Range` attributes in `*Current` modes.
@@ -41,7 +39,7 @@ namespace PokitPro {
         AutoRange = 255 ///< Auto-range.
     };
     QTPOKIT_EXPORT QString toString(const CurrentRange &range);
-    QTPOKIT_EXPORT QVariant maxValue(const CurrentRange &range);
+    QTPOKIT_EXPORT quint32 maxValue(const CurrentRange &range);
 
     /// Values supported by the Pokit Pro's `Range` attributes in `Resistance` mode.
     enum class ResistanceRange : quint8 {
@@ -59,7 +57,7 @@ namespace PokitPro {
         AutoRange = 255 ///< Auto-range.
     };
     QTPOKIT_EXPORT QString toString(const ResistanceRange &range);
-    QTPOKIT_EXPORT QVariant maxValue(const ResistanceRange &range);
+    QTPOKIT_EXPORT quint32 maxValue(const ResistanceRange &range);
 
     /// Values supported by the Pokit Pro's `Range` attributes in `*Voltage` modes.
     enum class VoltageRange : quint8 {
@@ -74,7 +72,7 @@ namespace PokitPro {
         AutoRange = 255 ///< Auto-range.
     };
     QTPOKIT_EXPORT QString toString(const VoltageRange &range);
-    QTPOKIT_EXPORT QVariant maxValue(const VoltageRange &range);
+    QTPOKIT_EXPORT quint32 maxValue(const VoltageRange &range);
 
 }