+++ src/lib/abstractpokitservice_p.h
@@ -16,8 +16,6 @@
 #include <QLowEnergyService>
 #include <QObject>
 
-#include <optional>
-
 class QLowEnergyController;
 
 QTPOKIT_BEGIN_NAMESPACE