+++ src/lib/dsoservice.cpp
@@ -58,11 +58,9 @@ QString DsoService::toString(const quint8 range, const Mode mode) const
 }
 
 /*!
- *  Returns the maximum value for \a range, or the string "Auto".
- *
- *  If \a range is not a known valid enumeration value for \a product's \a mode, then a null QVariant is returned.
+ * Returns the maximum value for \a range, or 0 if \a range is not a known value for \a product's \a mode.
  */
-QVariant DsoService::maxValue(const PokitProduct product, const quint8 range, const Mode mode)
+quint32 DsoService::maxValue(const PokitProduct product, const quint8 range, const Mode mode)
 {
     switch (mode) {
     case Mode::Idle:
@@ -74,16 +72,13 @@ QVariant DsoService::maxValue(const PokitProduct product, const quint8 range, co
     case Mode::AcCurrent:
         return CurrentRange::maxValue(product, range);
     }
-    return QVariant();
+    return 0;
 }
 
 /*!
- *  Returns the maximum value for \a range, or the string "Auto".
- *
- *  If \a range is not a known valid enumeration value for the current \a product's \a mode,
- *  then a null QVariant is returned.
+ * Returns the maximum value for \a range, or 0 \a range is not a known value for the current \a product's \a mode.
  */
-QVariant DsoService::maxValue(const quint8 range, const Mode mode) const
+quint32 DsoService::maxValue(const quint8 range, const Mode mode) const
 {
     return maxValue(*pokitProduct(), range, mode);
 }