+++ src/lib/multimeterservice.cpp
@@ -83,11 +83,9 @@ QString MultimeterService::toString(const quint8 range, const Mode mode) const
 }
 
 /*!
- *  Returns the maximum value for \a range, or the string "Auto".
- *
- *  If \a range is not a known valid enumeration value for \a product's \a mode, then a null QVariant is returned.
+ * Returns the maximum value for \a range, or 0 if \a range is not a known value for \a product's \a mode.
  */
-QVariant MultimeterService::maxValue(const PokitProduct product, const quint8 range, const Mode mode)
+quint32 MultimeterService::maxValue(const PokitProduct product, const quint8 range, const Mode mode)
 {
     switch (mode) {
     case Mode::Idle:
@@ -109,16 +107,13 @@ QVariant MultimeterService::maxValue(const PokitProduct product, const quint8 ra
     case Mode::ExternalTemperature:
         break;
     }
-    return QVariant();
+    return 0;
 }
 
 /*!
- *  Returns the maximum value for \a range, or the string "Auto".
- *
- *  If \a range is not a known valid enumeration value for the current \a product's \a mode,
- *  then a null QVariant is returned.
+ * Returns the maximum value for \a range, or 0 \a range is not a known value for the current \a product's \a mode.
  */
-QVariant MultimeterService::maxValue(const quint8 range, const Mode mode) const
+quint32 MultimeterService::maxValue(const quint8 range, const Mode mode) const
 {
     return maxValue(*pokitProduct(), range, mode);
 }